+++ mockserver-core/src/main/java/org/mockserver/serialization/serializers/request/HttpRequestDTOSerializer.java
@@ -51,6 +51,9 @@ public class HttpRequestDTOSerializer extends StdSerializer<HttpRequestDTO> {
         if (httpRequest.getSocketAddress() != null) {
             jgen.writeObjectField("socketAddress", httpRequest.getSocketAddress());
         }
+        if (isNotBlank(httpRequest.getRemoteAddress())) {
+            jgen.writeObjectField("remoteAddress", httpRequest.getRemoteAddress());
+        }
         if (httpRequest.getBody() != null) {
             jgen.writeObjectField("body", httpRequest.getBody());
         }