+++ mockserver-core/src/main/java/org/mockserver/serialization/serializers/request/HttpRequestSerializer.java
@@ -51,6 +51,9 @@ public class HttpRequestSerializer extends StdSerializer<HttpRequest> {
         if (httpRequest.getSocketAddress() != null) {
             jgen.writeObjectField("socketAddress", httpRequest.getSocketAddress());
         }
+        if (isNotBlank(httpRequest.getRemoteAddress())) {
+            jgen.writeObjectField("remoteAddress", httpRequest.getRemoteAddress());
+        }
         if (httpRequest.getBody() != null && isNotBlank(String.valueOf(httpRequest.getBody().getValue()))) {
             jgen.writeObjectField("body", httpRequest.getBody());
         }